+++ src/main/java/cn/nukkit/network/protocol/EntityPickRequestPacket.java
@@ -8,7 +8,7 @@ public class EntityPickRequestPacket extends DataPacket {
     public static final byte NETWORK_ID = ProtocolInfo.ENTITY_PICK_REQUEST_PACKET;
 
     public long entityRuntimeId;
-    public byte hotbarSlot;
+    public int hotbarSlot;
 
     @Override
     public byte pid() {
@@ -18,13 +18,13 @@ public class EntityPickRequestPacket extends DataPacket {
     @Override
     public void decode() {
         this.entityRuntimeId = this.getEntityRuntimeId();
-        this.hotbarSlot = (byte) this.getByte();
+        this.hotbarSlot = this.getByte();
     }
 
     @Override
     public void encode() {
         this.reset();
         this.putEntityRuntimeId(this.entityRuntimeId);
-        this.putByte(this.hotbarSlot);
+        this.putByte((byte) this.hotbarSlot);
     }
 }
