+++ src/main/java/cn/nukkit/network/protocol/InteractPacket.java
@@ -10,12 +10,12 @@ public class InteractPacket extends DataPacket {
 
     public static final byte NETWORK_ID = ProtocolInfo.INTERACT_PACKET;
 
-    public static final byte ACTION_VEHICLE_EXIT = 3;
-    public static final byte ACTION_MOUSEOVER = 4;
-    public static final byte ACTION_OPEN_NPC = 5;
-    public static final byte ACTION_OPEN_INVENTORY = 6;
+    public static final int ACTION_VEHICLE_EXIT = 3;
+    public static final int ACTION_MOUSEOVER = 4;
+    public static final int ACTION_OPEN_NPC = 5;
+    public static final int ACTION_OPEN_INVENTORY = 6;
 
-    public byte action;
+    public int action;
     public long entityRuntimeId;
     public float x;
     public float y;
@@ -28,7 +28,7 @@ public class InteractPacket extends DataPacket {
 
     @Override
     public void decode() {
-        this.action = (byte) this.getByte();
+        this.action = this.getByte();
         this.entityRuntimeId = this.getEntityRuntimeId();
         if (this.action == ACTION_MOUSEOVER) {
             this.x = this.getLFloat();
@@ -40,7 +40,7 @@ public class InteractPacket extends DataPacket {
     @Override
     public void encode() {
         this.reset();
-        this.putByte(this.action);
+        this.putByte((byte) this.action);
         this.putEntityRuntimeId(this.entityRuntimeId);
         if (this.action == ACTION_MOUSEOVER) {
             this.putLFloat(this.x);
