+++ src/main/java/cn/nukkit/network/protocol/MobEffectPacket.java
@@ -11,12 +11,12 @@ public class MobEffectPacket extends DataPacket {
 
     public static final byte NETWORK_ID = ProtocolInfo.MOB_EFFECT_PACKET;
 
-    public static final byte EVENT_ADD = 1;
-    public static final byte EVENT_MODIFY = 2;
-    public static final byte EVENT_REMOVE = 3;
+    public static final int EVENT_ADD = 1;
+    public static final int EVENT_MODIFY = 2;
+    public static final int EVENT_REMOVE = 3;
 
     public long entityRuntimeId;
-    public byte event;
+    public int event;
     public int effectId;
     public int amplifier;
     public boolean enableParticles = true;
@@ -30,7 +30,7 @@ public class MobEffectPacket extends DataPacket {
     @Override
     public void decode() {
         this.entityRuntimeId = this.getEntityRuntimeId();
-        this.event = (byte) this.getByte();
+        this.event = this.getByte();
         this.effectId = this.getVarInt();
         this.amplifier = this.getVarInt();
         this.enableParticles = this.getBoolean();
@@ -41,7 +41,7 @@ public class MobEffectPacket extends DataPacket {
     public void encode() {
         this.reset();
         this.putEntityRuntimeId(this.entityRuntimeId);
-        this.putByte(this.event);
+        this.putByte((byte) this.event);
         this.putVarInt(this.effectId);
         this.putVarInt(this.amplifier);
         this.putBoolean(this.enableParticles);
