+++ src/main/java/cn/nukkit/network/protocol/PlayerHotbarPacket.java
@@ -9,7 +9,7 @@ public class PlayerHotbarPacket extends DataPacket {
     public static final byte NETWORK_ID = ProtocolInfo.PLAYER_HOTBAR_PACKET;
 
     public int selectedHotbarSlot;
-    public byte windowId = ContainerIds.INVENTORY;
+    public int windowId = ContainerIds.INVENTORY;
     public boolean selectHotbarSlot = true;
 
     @Override
@@ -20,7 +20,7 @@ public class PlayerHotbarPacket extends DataPacket {
     @Override
     public void decode() {
         this.selectedHotbarSlot = (int) this.getUnsignedVarInt();
-        this.windowId = (byte) this.getByte();
+        this.windowId = this.getByte();
         this.selectHotbarSlot = this.getBoolean();
     }
 
@@ -28,7 +28,7 @@ public class PlayerHotbarPacket extends DataPacket {
     public void encode() {
         this.reset();
         this.putUnsignedVarInt(this.selectedHotbarSlot);
-        this.putByte(this.windowId);
+        this.putByte((byte) this.windowId);
         this.putBoolean(this.selectHotbarSlot);
     }
 }
