+++ src/main/java/cn/nukkit/network/protocol/UpdateEquipmentPacket.java
@@ -8,8 +8,8 @@ public class UpdateEquipmentPacket extends DataPacket {
 
     public static final byte NETWORK_ID = ProtocolInfo.UPDATE_EQUIPMENT_PACKET;
 
-    public byte windowId;
-    public byte windowType;
+    public int windowId;
+    public int windowType;
     public int windowSlotCount;
     public long entityUniqueId;
     public CompoundTag namedTag;
@@ -21,8 +21,8 @@ public class UpdateEquipmentPacket extends DataPacket {
 
     @Override
     public void decode() {
-        this.windowId = (byte) this.getByte();
-        this.windowType = (byte) this.getByte();
+        this.windowId = this.getByte();
+        this.windowType = this.getByte();
         this.windowSlotCount = this.getVarInt();
         this.entityUniqueId = this.getEntityUniqueId();
         this.namedTag = this.getCompoundTag();
@@ -31,8 +31,8 @@ public class UpdateEquipmentPacket extends DataPacket {
     @Override
     public void encode() {
         this.reset();
-        this.putByte(this.windowId);
-        this.putByte(this.windowType);
+        this.putByte((byte) this.windowId);
+        this.putByte((byte) this.windowType);
         this.putVarInt(this.windowSlotCount);
         this.putEntityUniqueId(this.entityUniqueId);
         this.putCompoundTag(this.namedTag);
